requirements.txt
python-telegram-bot[job-queue]
google-generativeai
gspread
google-auth-oauthlib
geopy
python-dotenv
pandas
openpyxl